ci: reduce CI runs by disabling CI on push to '*'
While we still want to have CI after a merge on main, blindly
running CI on all branch pushes triggers to often unnecessarily.
Therefore we reduce it to our default branch, in case one wants to check
before a PR, it is possible to manually trigger a run on a branch.

Signed-off-by: Stefan Kraus <dev@stefankraus.org>